Fixed bug in mark function. Working simulations for bot cars and motorcycles.

+++ source/hdf_save.cpp
@@ -1,6 +1,6 @@
 #include "hdf_save.h"
 
-int hd5data(vector<vector<vector<short>>> data, float density, float car_ratio, int trial, char* _filename,
+int hd5data(vector<vector<vector<int>>> data, float density, float car_ratio, int trial, char* _filename,
 			long seed)
 {
     #define RANK 3
